+++ test/archived_form_elements_test.rb
@@ -10,6 +10,12 @@ class ArchivedFormElementsTest < Minitest::Test
     end
   end
 
+  class ModelWithArchivedTree
+    def inline_forms_attribute_list
+      [[:children, "Children", :tree]]
+    end
+  end
+
   def test_raises_when_model_declares_archived_geo_code_curacao
     err = assert_raises(InlineForms::ArchivedFormElementError) do
       InlineForms.validate_no_archived_form_elements_for!(ModelWithArchivedElement)
@@ -17,6 +23,14 @@ class ArchivedFormElementsTest < Minitest::Test
     assert_includes err.message, "geo_code_curacao"
     assert_includes err.message, "7.6.0"
     assert_includes err.message, "archived/form_elements/geo_code_curacao"
+  end
+
+  def test_raises_when_model_declares_archived_tree
+    err = assert_raises(InlineForms::ArchivedFormElementError) do
+      InlineForms.validate_no_archived_form_elements_for!(ModelWithArchivedTree)
+    end
+    assert_includes err.message, "tree"
+    assert_includes err.message, "archived/form_elements/tree"
   end
 
   def test_archived_registry_documents_absence_list_without_path